--- src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,11 +1,0 @@
-package kitchenpos.menu.domain;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-	long countByIdIn(List<Long> menuIds);
-
-	List<Menu> findAllByIdIn(List<Long> menuIds);
-}
