+++ src/main/java/kitchenpos/menu/domain/menu/MenuGroupRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.menu.domain.menu;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+}
