+++ src/main/java/kitchenpos/menu/domain/menu/MenuRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.menu.domain.menu;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+	long countByIdIn(List<Long> menuIds);
+
+	List<Menu> findAllByIdIn(List<Long> menuIds);
+}
