+++ src/main/java/kitchenpos/menu/domain/product/ProductRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.menu.domain.product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
