+++ src/main/java/kitchenpos/menu/dto/product/ProductRequest.java
@@ -1,0 +1,36 @@
+package kitchenpos.menu.dto.product;
+
+import kitchenpos.menu.domain.product.Product;
+
+import java.math.BigDecimal;
+
+public class ProductRequest {
+
+    private String name;
+    private BigDecimal price;
+
+    public ProductRequest(final String name, final BigDecimal price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public Product toProduct() {
+        return new Product(name, price);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    public void setPrice(final BigDecimal price) {
+        this.price = price;
+    }
+}
