+++ src/main/java/kitchenpos/menu/ui/menu/MenuRestController.java
@@ -1,0 +1,39 @@
+package kitchenpos.menu.ui.menu;
+
+import java.net.URI;
+import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import kitchenpos.menu.application.menu.MenuService;
+import kitchenpos.menu.dto.menu.MenuRequest;
+import kitchenpos.menu.dto.menu.MenuResponse;
+
+@RestController
+public class MenuRestController {
+    private final MenuService menuService;
+
+    public MenuRestController(final MenuService menuService) {
+        this.menuService = menuService;
+    }
+
+    @PostMapping("/api/menus")
+    public ResponseEntity<MenuResponse> create(@RequestBody final MenuRequest menuRequest) {
+        final MenuResponse created = menuService.create(menuRequest);
+        final URI uri = URI.create("/api/menus/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/menus")
+    public ResponseEntity<List<MenuResponse>> list() {
+        return ResponseEntity.ok()
+                .body(menuService.findAll())
+                ;
+    }
+}
