--- src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,30 +1,0 @@
-package kitchenpos.menugroup.application;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kitchenpos.menugroup.domain.MenuGroupRepository;
-import kitchenpos.menugroup.dto.MenuGroupRequest;
-import kitchenpos.menugroup.dto.MenuGroupResponse;
-
-@Transactional
-@Service
-public class MenuGroupService {
-    private final MenuGroupRepository menuGroupRepository;
-
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
-        this.menuGroupRepository = menuGroupRepository;
-    }
-
-    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
-        return MenuGroupResponse.of(menuGroupRepository.save(menuGroupRequest.toMenuGroup()));
-    }
-
-    @Transactional(readOnly = true)
-    public List<MenuGroupResponse> findAll() {
-        return MenuGroupResponse.ofList(menuGroupRepository.findAll());
-    }
-
-}
