--- src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,31 +1,0 @@
-package kitchenpos.product.application;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.dto.ProductResponse;
-
-@Service
-public class ProductService {
-    private final ProductRepository productRepository;
-
-    public ProductService(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    @Transactional
-    public ProductResponse create(final ProductRequest productRequest) {
-        Product product = productRequest.toProduct();
-        return ProductResponse.of(productRepository.save(product));
-    }
-
-    public List<ProductResponse> list() {
-        return ProductResponse.ofList(productRepository.findAll());
-    }
-
-}
