--- src/main/java/kitchenpos/product/domain/Price.java
@@ -1,31 +1,0 @@
-package kitchenpos.product.domain;
-
-import javax.persistence.Embeddable;
-import java.math.BigDecimal;
-import java.util.Objects;
-
-@Embeddable
-public class Price {
-
-    private BigDecimal price;
-
-    public Price(BigDecimal price) {
-        validate(price);
-        this.price = price;
-    }
-
-    protected Price() {
-    }
-
-    private void validate(BigDecimal price) {
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public BigDecimal price() {
-        return this.price;
-    }
-
-}
-
