--- src/main/java/kitchenpos/product/domain/Product.java
@@ -1,56 +1,0 @@
-package kitchenpos.product.domain;
-
-import javax.persistence.*;
-import java.math.BigDecimal;
-import java.util.Objects;
-
-@Entity
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    @Embedded
-    private Price price;
-
-    protected Product() {
-    }
-
-    public Product(Long id, String name, BigDecimal price) {
-        this(name, price);
-        this.id = id;
-    }
-
-    public Product(String name, BigDecimal price) {
-        this.name = name;
-        this.price = new Price(price);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price.price();
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        final Product product = (Product)o;
-        return Objects.equals(id, product.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
