--- src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,6 +1,0 @@
-package kitchenpos.product.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
