--- src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,36 +1,0 @@
-package kitchenpos.product.dto;
-
-import kitchenpos.product.domain.Product;
-
-import java.math.BigDecimal;
-
-public class ProductRequest {
-
-    private String name;
-    private BigDecimal price;
-
-    public ProductRequest(final String name, final BigDecimal price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public Product toProduct() {
-        return new Product(name, price);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
-    }
-}
