--- src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,43 +1,0 @@
-package kitchenpos.product.dto;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import kitchenpos.product.domain.Product;
-
-public class ProductResponse {
-
-    private final Long id;
-    private final String name;
-    private final BigDecimal price;
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public ProductResponse(Long id, String name, BigDecimal price) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-    }
-
-    public static ProductResponse of(final Product product) {
-        return new ProductResponse(product.getId(), product.getName(), product.getPrice());
-    }
-
-    public static List<ProductResponse> ofList(final List<Product> products) {
-        return products.stream()
-            .map(ProductResponse::of)
-            .collect(Collectors.toList());
-    }
-
-}
